Allows to simulate I/O errors
Introduces the "instrumentation" feature.

When enabled the "set_number_of_allowed_io_operations" is exposed and allows to set the number of allowed I/O operations.
After this threshold is reached all I/O operations returns an error.

This required to wrap all I/O calls with the new "try_io!" macro that is an alias for "try!" if the "instrumentation" feature is not enabled.